Make non-updatable stubs depend on sdk 29
There are a few classes from modules referenced as type
arguments in the non-updatable part of the API, which
need to be resolved.

The current dependencies that satisfy those references
create dependency cycles when we change the big stubs to
be assembled from other stubs. I considered two options for
breaking that cycle:
1) depending on a previous SDK
2) add empty classes in a simple manual stub lib

This CL does #1. #2 is more work, but may be required when/if
we want to reference a newly-added API in the same release.

Bug: 144149403
Bug: 149906971
Test: m
Change-Id: I3c5e611ca45536bdd931291e663a21757e35992f
Merged-In: I3c5e611ca45536bdd931291e663a21757e35992f
(cherry picked from commit d2ccda37e546bf0c802bafbcc82edb5dbd345ee3)