Disable activity process config override for certain display-based services.
Single session IME services expect their config to be overriden with the
display the IME window is currently on while launching an activity on a
different display will trigger the compat logic to override the config
with the recent display.

A11y and voice interaction services can also have their display config
overriden with activity launches.

Test: WindowProcessControllerTests
Test: MockingOomAdjusterTests
Test: ProcessRecordTests
Test: ActivityManagerServiceTest
Test: OomAdjusterTests

Fixes: 153843194

Change-Id: Ibe0dd25d11935d07362dcfe95f338deda6d6e3da