Respect parent bounds if it can contain the override bounds
The parent configuration may have been transformed (fixed rotation),
in that case the override bounds should be able to fit in the given
parent bounds. Besides, the rotation of task display area is different
than the override bounds. So parent app bounds should be still chosen
as the container area.

Fixes: 154699350
Test: atest AspectRatioTests#testMinAspectLandscapeActivity
            TaskRecordTests#testComputeConfigResourceOverrides
Change-Id: Iaf5c1f838d2c51c6c0175b5e6e947e6690ed9cf3