Update TvSettingsUIInteracted atom
This CL removes the string field from the TvSettingsUIInteracted atom to
eliminate the risk of logging sensitive information by accident. This
change should be safe as no such log has been uploaded to Google server
yet.

Besides the change to the atom, this CL also made some corresponding
change to the enum file so fill in the information that was collected
using a string.
(This CL is a mirror of cl/310419842 to Android repo.)

Metrics council review bug: 151768952
Bug: 150979527
Test: the Android tree still builds normally

Change-Id: I8a9d8c95ec06eaa2cbe097efe652e18d0e7baeac